Moved core directory.
