+++ package.json
@@ -24,8 +24,11 @@
   "scripts": {
     "lint": "eslint src",
     "rollup": "rollup -c",
-    "uglify": "uglifyjs dist/io.js -cmo dist/io.js --comments && uglifyjs dist/io-extras.js -cmo dist/io-extras.js --comments",
-    "build": "npm run lint && npm run rollup && npm run uglify",
+    "uglify": "uglifyjs dist/io.js -cmo dist/io.js --comments",
+    "uglify-elements": "uglifyjs dist/io-elements.js -cmo dist/io-elements.js --comments",
+    "uglify-extras": "uglifyjs dist/io-extras.js -cmo dist/io-extras.js --comments",
+    "uglify-all": "npm run uglify && npm run uglify-elements && npm run uglify-extras",
+    "build": "npm run lint && npm run rollup && npm run uglify-all",
     "travis": "npm run build"
   },
   "repository": {
