+++ src/elements/selector-sidebar.js
@@ -1,7 +1,7 @@
-import {html} from "../core/element.js";
+import {html} from "../io.js";
 import {IoSelector} from "./selector.js";
 import "./sidebar.js";
-import {filterObject} from "../core/utils.js";
+import {filterObject} from "../io.js";
 
 export class IoSelectorSidebar extends IoSelector {
   static get Style() {
