+++ src/elements/selector.js
@@ -1,4 +1,4 @@
-import {IoElement, html} from "../core/element.js";
+import {IoElement, html} from "../io.js";
 
 const importedPaths = {};
